Refactor Pipeline logic into it's own file
Moves all logic related to Pipeline into a new file (pipeline.rs).
