+++ src/lib.rs
@@ -359,13 +359,15 @@ assert_eq!(result, Ok(("foo".to_string(), b"bar".to_vec())));
 
 // public api
 pub use crate::client::Client;
-pub use crate::cmd::{cmd, pack_command, pipe, Arg, Cmd, Iter, Pipeline};
+pub use crate::cmd::{cmd, pack_command, pipe, Arg, Cmd, Iter};
 pub use crate::commands::{Commands, ControlFlow, LposOptions, PubSubCommands};
 pub use crate::connection::{
     parse_redis_url, transaction, Connection, ConnectionAddr, ConnectionInfo, ConnectionLike,
     IntoConnectionInfo, Msg, PubSub,
 };
 pub use crate::parser::{parse_redis_value, Parser};
+pub use crate::pipeline::Pipeline;
+
 #[cfg(feature = "script")]
 #[cfg_attr(docsrs, doc(cfg(feature = "script")))]
 pub use crate::script::{Script, ScriptInvocation};
@@ -431,5 +433,6 @@ mod cmd;
 mod commands;
 mod connection;
 mod parser;
+mod pipeline;
 mod script;
 mod types;
